🤔 rethinking render, refresh, reset and stripWidgets interface
